requirements.txt
python-dotenv
requests
PyGithub
GitPython
docker
